+++ mobile/android/android-components/components/browser/engine-gecko-beta/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineTest.kt
@@ -9,7 +9,6 @@ import android.content.Context
 import mozilla.components.concept.engine.DefaultSettings
 import mozilla.components.concept.engine.EngineSession.TrackingProtectionPolicy
 import mozilla.components.concept.engine.UnsupportedSettingException
-import mozilla.components.concept.engine.webextension.WebExtension
 import mozilla.components.support.test.any
 import mozilla.components.support.test.argumentCaptor
 import mozilla.components.support.test.mock
@@ -174,7 +173,8 @@ class GeckoEngineTest {
 
         `when`(runtime.registerWebExtension(any())).thenReturn(result)
         engine.installWebExtension(
-                WebExtension("test-webext", "resource://android/assets/extensions/test"),
+                "test-webext",
+                "resource://android/assets/extensions/test",
                 onSuccess = { onSuccessCalled = true },
                 onError = { _, _ -> onErrorCalled = true }
         )
@@ -198,7 +198,7 @@ class GeckoEngineTest {
 
         var throwable: Throwable? = null
         `when`(runtime.registerWebExtension(any())).thenReturn(result)
-        engine.installWebExtension(WebExtension("test-webext-error", "resource://android/assets/extensions/error")) { _, e ->
+        engine.installWebExtension("test-webext-error", "resource://android/assets/extensions/error") { _, e ->
             onErrorCalled = true
             throwable = e
         }